Implement way to conveniently use cmd args to make
The problem is that GNU Make doesn't have a way to reference arguments
provided by the user inside the targets. E.g., as I understand, this:

make docker-container-copy-from PATH-IN-CONTAINER PATH-IN-HOST

is not possible. There is no way to use the values inside the
docker-container-copy-from target.

The only possible way is to use variables, like:

make PATH_IN_CONTAINER=PATH-IN-CONTAINER PATH_IN_HOST=PATH-IN-HOST \
  docker-container-copy-from

which is very inconvenient, to type all those upper case variables etc.

So now it's been solved, acceptably in my opinion, by using very short
names for the variables and have a dedicated target, which will list all
such variables, their explanations and examples of use.