Add doc dir and rules for development
It's been decided to have git controlled files where various rules for
development, end use of the project will be present. This is because
even at this stage it's become hard to remember some details, e.g. in
this case it was 'when to increment the docker image tag'.

Currently those docs are not supposed to be very formal. But later they
will become the basis for the more formal documentation both for the end
user and developer(s)