Enable server/main.js restart on file change
Had to install the nodemon npm package for it. The idea is that there
will be need to modify the server file, which is used to serve the ecma
modules of the x-charts library for testing purposes. And it will be
convenient to restart the server in case its file (or files, if it
becomes bigger, than it's now) changed.