Start amending server main.js
Implement test dir list JSON response.

The idea is to implement functionality, which would allow to have e2e
test pages, where actual use cases could be tested in the web browser,
to be generated more dynamically, instead of just having an html file.

This should allow to share certain common stuff and if later there is a
need to change that HTML, it could be done in 1 place, w/o having to
change hundreds of HTML files.

With this commit the changes have started to be implemented in general
and 1 particular change has been finished: return the list of test cases
for a directory.

I.e. if the user just opens the localhost:8080 page — then he (later,
not implemented currently) should be able to see the list of available
tests (or groups of tests). So that the user can click on a link and
either see a similar list 1 level down or open an actual e2e case. For
this to be implemented later now we return such a list if a request is
made to either localhost:8080/list-dir/ or
localhost:8080/test/e2e/cases/some/dir/list-dir/