Update nodemon setup
In order not to restart the server process on each file change, e.g.
src/foo.js, but only on changes in the server itself, I've changed the
nodemon setup. Now it's passed the --watch option in the npm script
command. The Containerfile has been updated accordingly.

Also now it's the --watch, not --legacy-watch, which is not needed and
may be was never needed. It probably was ignored in the previous setup.