Redirect to test root, if / requested
If the user opened /, this is supposed to show something useful, not a
404 etc., so that the user doesn't have to manually type a longer URL.

The simplest way is just to redirect to another handled URL, instead of
implementing the same handling logic for this special case (/).

B/c the 'same handling logic' is supposed to be something like:
- take the URL to determine the page title
- and so on

Even on that first step the / URL introduces additional if/else
complexity. But with the redirection — nothing special, except for an
additional HTTP request, which is not a problem.