Install npm module livereload
Haven't updated the image tag, because it seems to be unnecessary. It's
unlikely that the user will checkout this commit and rebuild the image.
No point in it.

Also will update the doc to mention such cases, that not on every change
to the package.json the image should be rebuilt.