Setup livereload
Make it watch changes in the ./test/e2e/cases/ and ./dist/ directories.
So that the developer can save time on manually switching to the browser
and reload it, when changes to tests or dist are made. The automatic
'make flow-build' won't be set up. This is up to the developer how to do
it.