Add .DEFAULT target to Makefile
Expected this to prevent cases when something unexpected was passed to
Make, e.g. as happened with me:
  make docker compose up -d
which resulted in huge output (and probably some actions with the
file system by Make). But setting .DEFAULT target hasn't helped with
this, it doesn't help with command arguments, like here -d.

And for simply unknown targets it outputs an error message and probably
does nothing to the target file.

But I decided to leave it.