Fix server/main.js handling file requests
Apparently Livereload for CSS files sends requests separately. I.e. for
JS and HTML files it reloads the whole page and everything works. But
for CSS files it loads only the changed file with an appended dummy get
parameter (to disable cache). Because the file path was in this case
something like ./real-file.css?getParameter=123123123 this exact file
didn't exist on the server. So just strip that part and use the
appropriate file name.