Amend doc/dev.md file
Specify explicitly that in case of the package.json file changes, the
image must be rebuilt (and its tag incremented) only if the changes
involved some changes in the node modules used.

I.e. if it's just an additional npm script added, no need to rebuild,
b/c the package.json file is mounted in the container anyway and the
change will be in effect w/o a rebuild.