Add way to flow-build single file
The intention is that later it may be inconvenient to rebuild everything
in certain cases (e.g. if we are running a unit test repeatedly on a
file save). So this is just a proof of concept that it's possible to
convert a single flow source file into a js file, which would actually
be tested.

Also in this commit tried to use GNUMake variables modification. I.e.
the user supplies a make call with a path and then that path is modified
by GNUMake and used in an actual shell command.

Also started to use a convention that some make targets will have a
*-help target, which would allow the user to quickly see some help
information useful for that target.