Implement initial XCharts wrapper
Including the shadowDom initialization.

In general, the concept is: the user will provide the library with a div
element to which a shadowDom will be attached.

In it there will be a top canvas (which is supposed to be used for
drawing of the whole picture, e.g. if the user is about to save the
charts as an image, the canvas is supposed to have the picture drawn
right before the image 'save as' dialog is open. Otherwise the canvas
will remain transparent. Also in order to prevent the user from
interacting directly with components and their parts unintentionally.
E.g. if a component has to use a div element and the user clicks on that
div, or selects a text in it etc., the user should not be able to tell
if it's a div or an image. The whole library stuff on the page should
look like an image to the user.

And there will be an inner container which will contain containers
provided to components. The components will be free to use their
container as they see fit (with certain restrictions though, e.g. not to
use id attributes, which may collide with other ids, not modify the DOM
outside their container etc.). It's hard to predict now all the
restrictions, so everything will be amended/reworked later, as it
gradually becomes more clear how the library is used.