Replace _* private methods with functions
For consistency, I've decided not to use them. If later it's needed to
have some of them as methods — will restore them. But currently I'm
opting to use functions. One benefit is that if the user types something
in the console, he will see suggestions of only not private stuff. Which
should give a more clear representation of how that 'something' is
supposed to be used, what it can do.

Private properties have remained though, b/c they are not shared between
instances and it would be more effort to implement a similar handling
for them, better to just store them on the instance.