Remove 'Drawing' from names of pie-3d types
Originally those types had the 'Drawing' word as part of their names.
Now, it's been decided to store some other data in the same types, e.g.
the absolute value of the piece of data represented by a slice of the
pie's, which is not needed by itself to draw the pie with the canvas 2d
API. Hence the name changes.