Remove suffixes from pie-3d types
Instead of expecting strings with suffixes add the suffixes to the
property name.

This should simplify the logic, the user will be responsible to provide
ready to use numbers. I.e. instead of '20px' it will be just 20. This
also should be more convenient to the user, b/c it's unlikely someone
would use values like '20px' elsewhere.