Move unit tests to ./src/test-unit/
There was a desire (for the sake of order) to store unit tests in the
./test/unit/ dir. But (since the unit tests files are supposed to be
Flow files, for convenience):
- in order for the IDE stuff to work
- and the absence of a way to have the ./src/.flowconfig file to be
  stored in the ./ dir (I don't remember exactly why, but there were
  problems with the Flow parser parsing files it was not supposed to,
  IIRC the files in the ./dist/ dir etc.)
there was an idea to use a symlink from ./test/unit to ./src/, this way
it would theoretically work, we only had to have another .flowconfig
file in the ./test/unit/ dir.

But it turned out that the symlinks were not supported on Windows. So it
was seconded that we store the unit tests in the ./src/ dir.

But then it was decided to store them in the ./src/test-unit/ directory,
seemed more neatly.