Implement test/served-tmp/ directory
The intention is to serve arbitrary files from it (currently mostly HTML
ones are expected), in order to facilitate testing (and may be for some
other purposes later).

E.g. if there is a need to inspect something visually, e.g. when a
console STDOUT output is not enough, like when you have lots of data,
which are supposed to be passed to the Canvas2d API, it may be easier to
draw what's needed with those data and see what's wrong.

In such a case the developer is supposed to write an HTML file in a unit
test and open that file in the browser.