Pre-calculate context2d.ellipse() arguments
It's been decided to have radiusX/Y, rotation values in this type. Those
are arguments for the context2d.ellipse() method. Originally I thought
that they are redundant, b/c these values could be calculated from the
points (center x, y, top x, y, left edge x, y). But then it turned out
that it would be convenient to have them pre-calculated, b/c in the
tests, when we use utils to draw something from the PieData, it would
require to make those calculations (where bugs could occur) in the test.
Not convenient.

Better to have them pre-calculated once, hard-coded in the tests and
then rely on that data. It will be unlikely that in the test data
(hardcoded) an error will occur.