Finish calculation for points before transforms
I.e. when the pie is just a circle, not rotated, its points now seem to
be positioned correctly.

If that turns out to be not true, may be the util drawDataOnCanvas()
will have to be amended to draw the circle itself. Currently, just a
line from the center to the right edge seems to be sufficient.