Implement new way for npm modules
As a test case, the nodemon npm module has been updated. New Makefile
targets have been added to update npm modules. The process is basically:

  npm outdated

Then run:

  npm i --save-dev some-module

For each module manually.

-----------------

Then new way was necessary b/c the old way relied on mounting the
package.json & package-lock.json files as volumes for the Podman
container. This proved to be problematic, b/c if the file was
removed/created anew on the host machine (which apparently happens, when
one runs 'git restore') 2 separate files existed from that point (on the
host and container), and the changes made to the container's file were
not visible on the host hence couldn't be committed etc.

Now if the developer installs/updates an npm module the changed file in
the container is copied to the ./var/ dir and the changes are visible.