Make changes in Makefile
Remove the 'make flow-build' target. Originally the idea was that it may
be useful to build just particular files, not all of them, e.g. for the
purpose of running a single e2e test.

But it's been decided to throw it away, mainly b/c for the e2e tests it
should be enough to build everything then run all the tests. If it turns
out that we need it (e.g. if it gets too slow to rebuild everything
every time, then we will do it similarly to the way the unit tests'
build is done).

As for the removed target it seemed never to work and I failed to make
it work with GNUMake & npm scripts (the 'npm run' ones)

Also moved some variables to the target scope (as it's explained in the
GNUMake docs).