Make prepareData() test pass
Still needs refactoring, bugs may be present, b/c only 1 case was
tested, potentially may be problems when the start angle is PI or 2 *
PI. Also there is a need for validation, that the startAtDeg value from
the user must be in range 0 <= startAtDeg < 2 * Math.PI. And so on.

The point is that now it works at least in 1 case and there is
everything required to promptly fix bugs, if any.