Remove Makefile target for auto creating directory
This directory can be missing, b/c it doesn't have any committed files.
And originally the problem was solved with creating that dedicated
target. But since then other directories with the same potential problem
appeared: test/served-tmp/, test/diff/. And their existence is checked
only when the container is started, which seems to be sufficient. So
it's been decided to remove the target and follow suit of the other 2
directories.