Use expected value data from stringify-any
I.e. before this commit it used to be formatted like:

someField: [ 234, 123,
  234, ]

Now it's been decided that there is no need for it. It takes time to
format it this way etc. Just use the writeToTestDiffDir() util, it will
write the expected/actual files. Take the correct value as is and use it
in the test. This would save time and it's formatted pretty normally,
except that the object fields are double quoted, which seem not to
violate Google TS style guide (didn't read in full but a text search for
'quote' hadn't find any mention, that using "quotedProp" in an object
literal is forbidden)