Rotate pie around its inner center
We used to rotate the pie around the center of its Heads side. This led
to the pie being shifted (for a demonstration one can uncomment the
drawDataOnCanvas() call in the committed test) and compare the new
expected data and the old one.

Now it's rotated around its inner, true center.