Introduce start/endAngleOnEllipseClockwise fields
The unit test for the prepareData() module has been modified to pass,
and skipped, it's sort of dummy temporarily.

Currently those are not calculated. They will be needed to draw the pie
as start / end angles arguments to the .ellipse() methods.

Those angles turned out to be slightly trickier than I thought. They
have values relative to the ellipse's axes, which are rotated along with
the ellipse itself. I.e. the OX, OY axes thru the ellipses center are
rotated counter clockwise by the rotation argument to the .ellipse()
method.

I'd thought that they remain in place and the start / end angle
arguments should be relative to those original unrotated axes.

All of that warranted the pre-calculation of the start / end angles in
the prepareData() function, which is being now implemented.