Make drawDataOnCanvas() util draw 'slice-wise'
I.e. one has now an ability to draw not the whole ellipse, but its
slices, by setting angleStartSliceIndex, angleEndSliceIndex values on
the argument to the util.

This will draw from start angle of the former to the end angle of the
latter. Should make it more convenient to test the values of the
PieData.slices[N].start/endAngleOnEllipseClockwise.