Set up drawing images in unit tests
The intention is to compare images with another npm module. Because
comparing only JS data is not convenient enough, so in some unit tests
(and later e2e tests) the images will be drawn and compared.

The dummy src/test-unit/components/pie-3d/canvas-sandbox.test.js test
fill will have to be removed later, currently let it stay for reference
if needed.