Implement wrapper for looksSame()
The idea is that the function has many options (which are error prone)
and it's error handling is not perfect. So this wrapper along with a
sandbox test (for demonstration) has been created. The test can be
removed later.

No expected images have been added yet (in order not to store dummy
stuff in the repository).