Add example with many file names to Makefile
The Makefile.current (which is ignored in this Git repo) has a
potential to have many file names to be Flow compiled in a single
target. This makes it unwieldy. It turned out that GNUMake allows to
have strings formatted like in the committed change, which will allow to
have 1 file name per line, way more convenient.