Perform flow build, test false pass
The src/test-unit/test-utils/compare-with-looks-same.util.js util
couldn't detect an error in the image. So will have to modify it and the
test. In the only test for the draw() module the difference is the
missing stroke like on the right edge, but the test passed, which is
wrong.