Fix compareWithLooksSame() util, make test fail
The problem is in the only test case for the pie-3 draw() module: there
is a missing stroke line on the right edge of the pie. The test had not
detected it, so I had to make changes to the compareWithLooksSame()
util. The changes are:

- by default the comparison is strict, should be possible to change by
  providing options to the compareWithLooksSame()
- the diff file will be saved if the test fails and the diff file name
  is provided, yet it's slightly inconvenient to see where is the
  problem on the image, b/c the looks-same npm module doesn't allow to
  save a transparent image, like e.g. pixelmatch does.

So this has been addressed with this approach:

  - once the test fails, the developer will look into it, probably will
    make it to be run with the .only() tape method call, and re-run the
    test.
  - then in the console there will be a message:
    'DIFF BOUNDS on image foo.png: { left: 123, top: 123 ...}'.

    So the developer should have an idea of where is the problem.