Fix failing test, perform Flow build
Now the development of the prepareRimSlicesData() module in its own
branch can be considered finished. Tests pass, everything looks correct.

Subsequent changes to the draw() and prepareRimSlicesData() (in cases
where the draw() doesn't work, e.g. when the pie is rotated 90 deg.
around the center X axis) are expected to be minor and can be made in a
different branch.