Fix problem in prepareData()
The problem occurred when the left edge lied exactly at the boundary
between 2 slices. Haven't made the same change for the right edge, b/c
the subsequent validation is supposed to forbid rotations >= Math.PI *
2.

Anyway if later the problem with the right edge reveals itself, then
will make the same change for the right edge, currently I'm just not
feeling comfortable to make it, w/o knowing for sure if it's needed.