Rename 2 types
Rename RimSliceData to RimSliceDataFaceVisible, the intention is to
introduce another type RimSliceDataFaceInvisible, which won't have
properties named like pointStartOnVisibleFace etc. and this new type
will be used in case when neither the face, nor tails are visible. I
couldn't use the RimSliceData type as is b/c it would be confusing,
unclean etc. to have such named properties, when actually no face is
visible.