Rename type
The idea is to rework the prepareRimSlicesData() so that it doesn't use
props like pointStartOnVisibleFace, but instead pointStartOnHeads etc.

This should allow us to handle the case when no pie face (neither heads
nor tails) is visible. Otherwise it would be inaccurate/confusing to
have something like (in pseudo code):

  if !isHeadsVisible && !isTailsVisible:
    line from pointStartOnVisibleFace to pointStartOnInvisibleFace

So this renaming is the first step in that direction