Rename type properties
Namely get rid of names like pointStartOnVisibleFace etc. Instead in the
types use pointStartOnHeads etc., but in the draw method where the logic
depends on what's visible, check that and take for instance
pointStartOnHeads or pointStartOnTails.

This is needed to later use the same data to draw the pie in case when
it's rotated 90⁰ over the centerX axis, when neither heads nor tails is
visible. So that we don't end up with confusing code like: take
pointStartOnVisibleFace when no face is visible.

The tests pass.