Refactor
This refactoring would've been justified in any case, because it seems
that it's significantly reduced code duplication, although it was done
as part of preparation for implementing the case, when neither the
heads, nor tails is visible, that is when the pie rim is just a bar,
rectangle.

Also in this case (when the rim is just a bar) we don't need the
ellipseArgumentsOnHeads, ellipseArgumentsOnTails but, it's been decided
to still generate them, although with dummy values. Because had we made
those values nullable, it would require us else where to add additional
checks for "if it's not null". So to leave them as is seemed to be more
convenient, easier to support, use etc.