Preliminary implement 'rim bar' drawing
It works, although in 1 case the border between 2 slices seems to be too
bold. As I started to look into it, I found that in this particular
setup, if a face is visible, then the pie is not drawn correctly.

So it's been decided to save it as is and look deeper later.