Modify edge indices logic in pie-3d prepareData()
The main problem was (and it's not been solved completely) that the
slice index of an edge (left/right) of the pie's were determined
incorrectly. The logic used to be: when moving over slices, have flags
on which edge is now expected to be passed and depending on angle values
of the slice being passed set the left/right edge's slice index.

Now it's different: determine angles of the left/right edge (which can
depend on the start angle value). E.g. if we start from angle 270 deg,
which is planned to be allowed, then our left edge is supposed to be 360
+ 180 deg.

It's hard to figure out all the possible cases, so the approach taken is
to fix/amend the logic gradually, as various non-working cases are
discovered.

Currently, I haven't checked if the draw() tests pass, etc. And even if
they pass, 100% the pie won't be correct in case the
.rotationAroundCenterXAxisDeg value is > 180 deg, 270 deg. B/c currently
only the case with > 90 deg is handled.