Add 2 tests
This commit is closely related to the previous one:

  commit b65125e91c551043823a12b33371e8a3748f70d5
  Refs: [implement-pie-3d-drawing]
  Author:     Dmitry Koroliov <d.koroliov@gmail.com>
  AuthorDate: Sat May 31 13:47:20 2025 +0300
  Commit:     Dmitry Koroliov <d.koroliov@gmail.com>
  CommitDate: Sat May 31 13:47:20 2025 +0300

      Modify edge indices logic in pie-3d prepareData()

Basically, b/c of the change in the mentioned commit the 2 tests passed
w/o any code change. 1 other test: 0006-start-at-deg-90-data-75-25 in
the pie-3d draw() failed, but the difference was 2 pixels. I couldn't
see the difference with my eyes. Also this added 0006 test was the one,
which looked correctly in the unit tests, but failed in browsers:
Chrome, FF, Edge. Now it is drawn correctly in the browsers. B/c of all
that it's been decided just to replace the png image for the 0006 test
with the new one.