Add 2 skipped tests
It's been decided to commit but skip those added tests, b/c in order to
make them pass changes must be done in the data calculation logic,
namely:

  - all angles should be consistent, i.e. we start count from angle 0 at
    the right edge, move counter clockwise
  - then we start at the start angle and every angle (start, end) of
    every slice should include that start angle.
  - also the edges should store their angles, e.g. Math.PI * 4 if we
    start at, say, 350 degrees and so on.

After all of that is done, I'll resume the work on the skipped tests.

The problem with the skipped tests is that the logic of which slices are
visible on the rim and hence to be calculated is incorrect. In the
skipped test the left edge slice is slice 0 and it's also takes the
right edge. This makes the current implementation to draw just 1 single
slice on the rim. The direct cause is in:

src/components/pie-3d/prepare-rim-slices-data.js:96
  function getIndicesToPassThru() {

but to solve it all the mentioned changes must be done.