Modify unit test util as part of types rework
This commit on its own will break the tests, at least the util won't
draw correctly. It's committed separately b/c I wanted to emphasize the
strange change (Math.PI * 6 - start/endAngleCounterClockwise) and
explain it more deeper, than a comment in code would've fit.

The types rework is to have in SliceData actual, intuitively clear
angles, starting from the right edge (0), counterclockwise, then +
startAngle provided by the user and so on.

Currently it's a bit messy, because when the types were designed it was
expected that PieData would be enough to draw the pie. Then it turned
out that RimSlicesData would be needed to and so on.

Long story short it turned out to be messy and inconvenient to deal with
slices, edges etc. So a decision has been made to have all angles
consistent, counterclockwise etc. But the .ellipse() method works with
angles clockwise, i.e. if one draws an ellipse from angle 5.5 to 7
counterclockwise it won't draw it on the short path, as one might
expect. It would draw it on the long path. Because regardless of the
isCounteClockwise argument to the .ellipse() method, it counts the
angles clockwise. So the 5.5 angle is on the top right quadrant and 7 is
on the bottom right.

Nevertheless, I decided to proceed with the standard, intuitive angle
measures (counterclockwise, from 0 (the right edge)).

But for the util to work, we had to convert the angle from that system,
to the one expected by the .ellipse() method. The easiest way to do it
seemed to be just take large enough angle and subtract from it the
counterclockwise angle. This way you'd get the .ellipse()-expected angle
for the same point you see. Example, for angle 7 counterclockwise (in
the top right quadrant), to get it's .ellipse()-expected value, take 6
PI - 7 = 4 PI - smth small (7 - 2 pi), which will be in the same
quadrant, the same point, but in the .ellipse()-expected measurement.

Why the 6 PI ? B/c the start angle provided by the user will be
constrained to the range [0; 2 PI). Plus the 2 PI of the slices to draw
means that an angle value is not supposed to reach 4 PI. To be sure,
since it's a util, just added + 2 PI just in case.

Currently the util works in 1 not yet committed test. If there are any
problems with it, will fix.