Fix pie-3d prepareData()
Previously the tests passed, the unit tests utils drew everything
correctly, but angles on slices were inconsistent, which was the
original intention of the work in the current branch. E.g. on, say,
slice[1] the end angle was 5, but on slice[2] the end angle was 0.5. Now
all of it was fixed. The other tests for the prepareRimSlicesData(),
draw() can fail, didn't check. Will do after this commit.