Introduce isTop/BottomRimVisibleToUser flags
It used to be just isRimVisibleToUser, but now, in the process of fixing
the 2 skipped tests, it's been realized that we need to separate them.
When we know which rim is visible, we will be able to find which slices
are visible on the rim, b/c it's not enough to just know which slices
take the left/right edges of the pie. See the skipped tests and the draw
util calls in them, to see a case for the problem.