Introduce fields to types in PieData
The intention is to make the skipped tests pass. The easiest & most
reasonable, convenient way to do it seems to be this:

Provided that we already have the pie to be drawn correctly when it's
rotated around the center-x axis in range [0, Pi], in order to make it
to be drawn correctly we just have to treat rotations in the range (Pi,
2 * Pi) as rotations in the former range, but with slices reversed and
all angles calculated as 2 * Pi - angle.

E.g. if the pie has start angle is 90⁰ (counter clockwise, as always,
unless it's stated otherwise), then when it's rotated, say, 200⁰ it's
the same pie as if:

- heads becomes tails and conversely
- its slices are all reversed (i.e. 0 becomes length -1, 1 becomes
  length - 2 etc.),
- the start angle (of the pie, the one provided in the type
  AddComponentPie3dArgument) is 2 * Pi - the original start angle
- for every slice every angle becomes 2 * Pi - angle, e.g. 300⁰ becomes
  360 - 300 = 60⁰
- for every slice the end angle becomes the start angle and vice versa
- the pie's rotation around the center-x axis is just 20⁰, not 200⁰.

Then just draw the new pie.

So in order to do it, these additional flags have been introduced.

This seemed as the best solution, b/c otherwise we would have to have
numerous if/else etc. checks in many places, where we draw. Currently we
seem to have only 1: if the rim is at the top of the visible face or not
(except of course if the rim is not visible at all). Otherwise it would
be if the rim is at the top or bottom, if it's the original top rim
(before the rotations) or not, if to draw clockwise/counterclockwise, if
angle is angle or 2 * Pi - angle etc.