Modify draw PieData util
Originally I implemented different changes to that util, which would
make it to draw the reversed PieData, but show everything in accordance
with the user's provided argument. I.e. slice 0 is drawn as slice 0,
its hints are h0 and t0, not t7 and h7 respectively.

But then I realized that the util had become too complicated, not
convenient. So I just reset the commit, didn't even want to preserve it
in the repository history.

This change is way simpler. In case the pie is reversed only noticeable
text is shown to the developer: Reversed. For the developer it will be
easier to figure out which dot on the canvas corresponds to which data
entry, then to deal with the util, if it doesn't work correctly etc.