Make 1 skipped test pass
The change has been done according to the description in the original
commit in this branch. Basically if the rotation angle around the center
X axis is >= 180⁰, then we just 'reverse' the slice and rely on the
fact? that currently the pie is drawn correctly, if that angle in the
[0, 180] range. Despite the pie was drawn correctly for the rotation
angle 180, for consistency we reverse the pie if it's 180 too.

So now the PieData is not reversed if the rotation angle is [0, 180) and
it is reversed, if it's [180, 360).