Rework types for components in general
It used to be just a Component interface and in the XCharts module and
its registerComponent() function we used the Class<Component> utility
type to refer to the type of a component instance's class.

Now it's different. We have the ComponentInstance interface (still
empty) and a ComponentClass object type. The latter causes problem when
this ComponentClass is called like new ComponentClass(), but this error
has been suppressed with a $FlowFixMe comment. The FlowFixMe was chosen
b/c it was suggested by the Flow LSP server's code completion prompt.
Also this suppression seems not to affect any further type checks, e.g.
if a property is added to the ComponentClass type and is missing in a
component class definition, then Flow will throw etc.

This approach with suppression has been chosen as the most optimal.
Otherwise we would either have to use the previous approach, with the
Class<> utility type on the ComponentInstance interface (which also had
suppressions) or we would have to perform some insane type checks.

The suppression in the src/components/pie-3d/main.js on the
registerComponent() call, the reason for it was that otherwise we'd have
to declare the validateAddComponentArgument() type in the class
declaration, then outside of it do:

Pie3d.validateAddComponentArgument = function

I didn't want to do it, who knows, may be it affects performance etc.