Rework registerComponent() mechanism
I've removed all the error checks from it. The justification is this:

Originally the intention was to allow some other external developers to
add their own components, so that's why the removed error was
introduced. But this turned out to be something, which would require a
way more deeper planning. So as of now it won't be supported and only
X-Charts's own components will be allowed.

So all such error cases, like 'component already registered', etc. are
internal errors and should not exist. My understanding of an internal
error is that it's something which helps save time on debugging, when
it's not feasible to eliminate the possibility of such an event (b/c of
some architecture problems etc.). As of now, there is no need for that
error, b/c it's not a very burdening responsibility for the developer to
pay attention to if the component type already exists etc.