Setup validation for pie-3d .add() argument, build
No checks have been added, but the unit tests have been setup.

Regarding the type case of AddComponentArgument to
AddComponentPie3dArgument, it's deemed appropriate to use Flow error
suppressions there. Otherwise we'd have to do some very complicated
property checks etc. The idea was that we accept the argument provided
by the user to the .add() method of the xCharts instance and during the
validation we treat is as AddComponentPie3dArgument, because in case of
.type === 'pie-3d' this is exactly the component which is supposed to be
instantiated. Seems very simple and clean, shouldn't lead to any
problems.

Also a Flow build has been done.