Setup general AddComponentArgument validation
Some properties on that argument, as well as the argument itself should
be validated separately, not by the component like pie-3d.

This turned out to be complicated enough to justify moving it to a
separate module. No tests have been created for the module yet.

The idea is that the added module will check the argument itself (e.g.
if it's an object etc.) and some common properties (currently only the
zIndex and type). Everything else will be checked by the component.

The component also will check for extra properties present, by taking
the all enumerable properties of the argument and also those, which have
been validated by the added module (it exports those properties).