Save progress before interface change to JSON
The decision has been made to rework the user facing interface, and make
the XCharts constructor as well as its instances methods to accept a
JSON string instead of objects.

Advantages of this is that we will greatly simplify the validation,
since we will have to bother only with presence/absence of properties
and the correctness of their values. No need to check where the value
came from (a prototype or an own property), no getters/setters, no
references in the argument, e.g. if data[0] is the same ref as data[3]
etc. This should greatly simplify the validation and reduce potential
use problems.