Don't export value
The intention is to rework the validation with a different approach,
based on Object.keys(). And now, a component's validator won't accept
the properties, which have been handled by the general .add() method
argument validator, but instead all the props, which have not been
handled and must be handled by the component's validator.