Introduce type ValidationMapperPure
This change is part of wider effort to implement the validation in a
clean, automated way, namely not to hard code the argument structure
in the validation function, but to make it work automatically with the
data structure of any depth.

I.e. as a validation mapper we will create an object, which would match
the expected structure of the argument. In the validation function we
will go thru that object and check every of its properties. If the prop
is a function, it will be called, else if it's an object the function
will go deeper and do the same on the new level, then it will go up.

This way in order to add/remove/change argument structure we will have
to modify only the mapper object. The validation function will work.