Make xCharts.add() type be mixed
The justification was this: we have types in order to help the
development, not some automatic generation of documentation.

So, on one hand it's misleading to type the argument of the .add()
method as 'mixed', because it expects a single argument of type
AddComponentArgument, which itself is a super type of some specific
component's constructor argument, e.g. AddComponentPie3dArgument.

On the other hand, when we look into the source code at the beginning of
the .add() method, there is no guarantee, that the argument[0] is of the
required type, b/c it's provided by the user and can be anything.

Also added tests to check the number of arguments to the .add() method.