Simplify validation error codes
It seems that it won't be possible to have a code for every case where
an error is thrown. The codes have already become rather unweildy.

On the other hand I don't won't to get rid of them, because it seems as
if they still can be useful.

So as of now, let them be this way. Later may be it becomes more clear
how to organized validation error messages.