Add check for empty array in validation
Currently the need is to enforce the .data prop to be a non empty array.
It's been done by enforcing all values, which have an array entry in the
validation mapper to be a non empty array.

Currently this should suffice. If later there is a need to allow empty
arrays, then the validation for the 'array entries' (like arg.data
currently) in the mapper can be changed. May be move the object to
element 1 (currently it's at element 0 in the mapper, and on element 0
put a validation function). Something like that.

Currently there is no need for this, it would be over-complication.