Amend validation message insertion in HTML
The messages as of now, contain special characters, like >, e.g.
B/c of this it's been decided to escape them this way. If later there is
a need for something else, e.g. have HTML in an error message, then this
approach will have to be changed. Currently no need for it.