Add 2 more tests
It's been decided to allow non-integer properties on arrays. It's not
our business what else is present.

On the other hand, it's something unexpected. and we throw in case it's
an unknown property in an object (as opposed to an array). But in case
of the object, it's very likely that the user has misspelled some
correct property, so it makes sense to throw.

In case of a non-integer property present in an array object, this is
something which can be done on purpose. So makes sense to allow it.