Wrap existing .add() arg validation in function
The idea is this: we need to validate the structure of the argument:

  - every property exists
  - no stray props exist (e.g. if there was a typo)
  - if a property's primitive value is correct or if it has an array or
    object as its value, then go down and repeat the logic.

This all have been done and rather flexible. If there is a need to add a
property to an argument, modify the type, modify the validation mapper.

But there may be other kind of validation required: right now it's to
check that the total value of all data entries is > 0. Later other cases
may be needed, when it will be required to check that, e.g. 2 levels
deep there is a foo prop with value 'a' and 3 steps right 1 level deep
there is a bar prop with some value etc.

So it's been decided to have 2 steps in this validation: the first one
is the one which already has been implemented (the one, wrapped in its
own function in this commit) and then any other more complicated steps,
and those validation steps will be able to rely on the argument's
structure to be correct etc.

This seems to be like a simple and clean solution. Otherwise we'd have
to devise something rather complicated and it's not obvious if there is
any need for it ever.

------------

If something more complicated is ever needed a possible way can be:

  - for array combined entries validation, e.g. if it's required that
    the total of all entries in the array equals something, then one can
    add an ArrayCombinerValidator which would provide methods to call on
    every element of the array, once the method from the validation
    mapper has been called. It will store somewhere all the needed data
    and on the final element will validate the combined value.
  - for objects, probably another mapper has to be created, which would
    also provide some methods, which would be called once the object has
    been validated by the main (existing) validation.

As I mentioned, it's not obvious if this is ever needed. The only
potential benefit visible now is that there won't be a need for yet
another array (the data one or ponentially a different one) traversal.
But even it seems to have little practical benefit, since the .data
array hardly can be expected to have over 20 entries.

------------

No changes have been done in this commit, except for wrapping the
existing code into another separate function (and wrapping long lines
lengths etc.)